+++ commutative_reducer.c
@@ -13,7 +13,7 @@
 extern __thread struct __cilkrts_worker *tls_worker;
 
 
-void eval_commutative_red(int* image) {
+void eval_commutative_red(uint8_t* image) {
   Hist* local_views;
   //local_views = malloc(CILK_NWORKERS * sizeof(Hist));
   local_views = aligned_alloc(64 * 8, CILK_NWORKERS * sizeof(Hist));
@@ -28,7 +28,7 @@ void eval_commutative_red(int* image) {
     cilk_for (int j = 0; j < PAR_ITER; j++) {
       //int worker_number = (int) (*(((uint64_t*) tls_worker) + 4));
       int worker_number = __cilkrts_get_worker_number();
-      int pixel = image[j];
+      uint8_t pixel = image[j];
       local_views[worker_number].ele[pixel]++;
     }
   }
@@ -37,7 +37,7 @@ void eval_commutative_red(int* image) {
   }
   fasttime_t stop = gettime();
 
-  printf("%f\t%d\t%d\t%d\n", tdiff_sec(start, stop), n.ele[0], n.ele[1], n.ele[255]);
+  printf("%f\t%u\t%u\t%u\n", tdiff_sec(start, stop), n.ele[0], n.ele[1], n.ele[255]);
 
   free(local_views);
 }
